New method - Grid
I choose to change the way the map maker works, instead i implemented a grid where shapes snap to cells and centered automatically.

I'm searching on how to fix the bug which when you zoom and draw, the drawing will be shown at the same place as if you didn't zoom. And when you zoom out the drawing zoom out so it become really tiny